refrator function get ressource

+++ src/backend/MySystemInfo.h
@@ -8,7 +8,7 @@
 
 
 struct MySysInfo{
-    int time{};
+    long time{};
     double totalVirtualMemory{};
     double usedVirtualMemory{};
     double physicalMemory{};
